media-feed-concept/requirements.txt
Flask==2.3.3
newspaper3k==0.2.8
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 